getFetchEndpoint and removal of `oauth` authType.
After another round of close inspection and an outage during the
deployment of gitspiegel, here are the findings:
1. `oauth` authType was deeply incorrect, as it's not really possible to
use it in the same way as other auth types.
2. Getting token for fetching wouldn't work with "app" authType.

Thus, removing `oauth` authType together with its parameters, and
allowing `getFetchEndpoint` only for `token` and `installation`
authTypes.

I tried to make it forbidden via static typing, however that proved
itself being a bit too complex, so runtime check it is then.

`getFetchEndpoint` implementation also contained a bug. It picked the
token from the parameters, instead of getting the token via GitHub auth.